feat: create report package
