+++ packages/report/index.ts
@@ -2,15 +2,15 @@ export { VerboseReporter } from './src/reporters/Verbose'
 export { EmptyReporter } from './src/reporters/Empty'
 export { Option } from './src/types/Option'
 export { expect } from './src/types/Expect'
+export { IReporter } from './src/runtime/IReporter'
 export {
-	IReporter,
 	MeasurementKind,
 	TraceData,
 	TestEvent,
 	CompoundMeasurement,
 	NetworkTraceData,
 	CompositeTraceData,
-} from './src/runtime/Reporter'
+} from './src/types/Report'
 export { TestScriptError, TestScriptErrorMapper } from './src/runtime/TestScriptError'
-export { Assertion } from './src/runtime/Assertion'
+export { Assertion } from './src/types/Assertion'
 export { SourceUnmapper } from './src/utils/SourceUnmapper'
