+++ packages/report/src/types/Report.ts
@@ -1,5 +1,4 @@
-import { Assertion } from '../runtime/Assertion'
-import { TestScriptError } from '../runtime/TestScriptError'
+import { Assertion } from './Assertion'
 
 export type MeasurementKind =
 	| 'throughput'
@@ -99,60 +98,4 @@ export enum TestEvent {
 export interface TestLifecycle {
 	label: string
 	event: TestEvent
-}
-
-export interface IReporter {
-	reset(stepName: string): void
-
-	responseCode: string
-	stepName: string
-
-	/**
-	 * Writes a measurement to the collection service
-	 *
-	 * @param {string} measurement
-	 * @param {string} value
-	 * @param {string} label
-	 * @param {string} responseCode
-	 * @returns {Promise<void>}
-	 * @memberof IReporter
-	 */
-	addMeasurement(measurement: MeasurementKind, value: string | number, label?: string): void
-
-	addCompoundMeasurement(
-		measurement: MeasurementKind,
-		value: CompoundMeasurement,
-		label: string,
-	): void
-
-	/**
-	 * Flushes measurements to the collection service
-	 */
-	flushMeasurements(): Promise<void>
-
-	/**
-	 * Writes a trace to the collection service
-	 *
-	 * @param {*} traceData
-	 * @param {string} label
-	 * @param {string} responseCode
-	 * @returns {Promise<void>}
-	 * @memberof IReporter
-	 */
-	addTrace(traceData: TraceData, label: string): void
-
-	testLifecycle(event: TestEvent, label: string, timeMeasurement?: number): void
-
-	testAssertionError(err: TestScriptError): void
-	testStepError(err: TestScriptError): void
-	testInternalError(message: string, err: Error): void
-
-	// NOTE that this is synchronous
-	// https://github.com/DefinitelyTyped/DefinitelyTyped/blob/master/types/node/v8/index.d.ts#L25
-	testScriptConsole(method: string, message?: any, ...optionalParams: any[]): void
-
-	// Used by EventEmitterReporter
-	addListener?(event: string | symbol, listener: (...args: any[]) => void): this
-	on?(event: string | symbol, listener: (...args: any[]) => void): this
-	once?(event: string | symbol, listener: (...args: any[]) => void): this
 }
