+++ packages/report/src/types/Status.ts
@@ -1,0 +1,1 @@
+export type Status = 'passed' | 'failed' | 'skipped' | 'unexecuted'
